skip error during indexing
When an error is returned by search3-java, we handle it by throwing a
generic bad_request exception. This is the error handling path for both
search and indexing requests. During indexing however, the exception is
not caught and it leads to a search_worker_manager process crash. The
process is restarted by the search3_sup supervisor and the indexing
begins again. This leads to an endless loop of trying to index the same
document if the documents causes some sort of error. We modify this
behavior by catching the thrown exception and ignoring the document.