Remove grpcbox and ctx dependencies and use gpb only
Search was bringing in grpc box and a bunch of other dependencies which are not
really used anymore.

Also cleaned up a bunch of config files used for grpcbox.

Needed to make sure to pass gpb the same compile options that grpbox was using,
which was transforming protobuf message names to lower case with an underscore.